small optimization to check/assert_field_class()
instead of subsetting the data.table and then retrieving the columns'
classes, we now get the classes first and then subset. this change
speeds the function up by a second or so, which is quite noticeable when
we're making function calls in downstream packages (e.g. gtfstools) that
should be quite fast